xblock-dance-discussion/requirements.txt
mysql-connector-python
-e .